fix the nginx config; hide allow origin header returned from backend
If any backend returns Access-Control-Allow-Origin we end up with a
header with multiple values and that's not allowed or makes sense so we
have to ensure we hide it.

Also, we need to the origin to an actual domain in order to send
credentials. Credentials are not allowed with wildcard.